# openapi.yaml
openapi: 3.0.3
info:
  title: Accounts Service API
  description: API для сервиса аккаунтов
  version: 1.0.0
tags:
  - name: Accounts
    description: Получение аккаунтов
paths:
  /account:
    get:
      tags:
        - Account
      summary: Получение аккаунта
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccount'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Account
      summary: Редактирование аккаунта
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAccount'
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccount'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /account/{login}/change:
    post:
      tags:
        - Account
      summary: Изменение счета пользователя login
      parameters:
        - name: login
          in: path
          required: true
          description: Уникальный логин пользователя
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAccount'
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccount'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /account/{login}/transfer:
    post:
      tags:
        - Account
      summary: Перевод на другой счет со счета пользователя login
      parameters:
        - name: login
          in: path
          required: true
          description: Уникальный логин пользователя
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAccount'
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccount'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /accounts:
    get:
      tags:
        - Account
      summary: Получение аккаунтов
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserAccountSmall'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /cash:
    post:
      tags:
        - Cash
      summary: Пополнение счёта или снятие денег со счёта
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CashRequest'
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: integer
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /transfer:
    post:
      tags:
        - Transfer
      summary: Перевод со счета текущего аккаунта на счет другого аккаунта
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: integer
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    UserAccount:
      type: object
      required:
        - login
        - name
      properties:
        login:
          type: string
        name:
          type: string
        birthdate:
          type: string
        sum:
          type: integer
    UserAccountSmall:
      type: object
      required:
        - login
        - name
      properties:
        login:
          type: string
        name:
          type: string

    CashRequest:
      type: object
      required:
        - sum
        - action
      properties:
        sum:
          type: integer
        action:
          type: string

    TransferRequest:
      type: object
      required:
        - sum
        - login
      properties:
        sum:
          type: integer
        login:
          type: string

    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Код ошибки
        message:
          type: string
          description: Описание ошибки